% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_changepoints.R
\name{find_changepoints}
\alias{find_changepoints}
\title{find_changepoints}
\usage{
find_changepoints(sequence, minimum_distance, process_count)
}
\arguments{
\item{sequence}{A vector of floats}

\item{minimum_distance}{The minimum distance between changepoints as percentage of sequence length}

\item{process_count}{The different amount of processes generating data in the sequence. For example if a sequence is generated by alternating processes like this: A B A B A B then the changepoint count might be 5 but the process count is 2}
}
\value{
The list of changepoints in increasing size
}
\description{
Returns the position of changepoints in the sequence
}
