% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_PyChest.R
\name{install_PyChest}
\alias{install_PyChest}
\title{install_PyChest}
\usage{
install_PyChest()
}
\description{
Initializes the package and installs/Updates PyChest into the local recticulate-Python evnironment
}
