% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_estimator.R
\name{list_estimator}
\alias{list_estimator}
\title{list_estimator}
\usage{
list_estimator(sequence, minimum_distance)
}
\arguments{
\item{sequence}{A vector of floats}

\item{minimum_distance}{The minimum distance between changepoints as percentage of sequence length}
}
\value{
The list of changepoints in order of likelihood
}
\description{
Returns the position of changepoints in the sequence
}
