% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_changepoints.R
\name{find_changepoints}
\alias{find_changepoints}
\title{find_changepoints}
\usage{
find_changepoints(sample, minimum_distance, process_count)
}
\arguments{
\item{sample}{A vector of floats corresponding to the piecewise stationary sample where the retrospective changes are to be sought}

\item{minimum_distance}{A real number between 0 and 1 corresponding to a lower-bound on the minimum normalized length of the stationary segments (as percentage of total sample length)}

\item{process_count}{The different number of distinct stationary processes present.}
}
\value{
The list of changepoints in increasing size
}
\description{
Returns the position of changepoints in the sequence. NOTE: PyChest needs to be installed first by calling `install_PyChest'.
}
\references{
\insertRef{khaleghi2014asymptotically}{RChest}

\insertRef{khaleghi2012locating}{RChest}
}
