% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_estimator.R
\name{list_estimator}
\alias{list_estimator}
\title{list_estimator}
\usage{
list_estimator(sample, minimum_distance)
}
\arguments{
\item{sample}{A vector of floats corresponding to the piecewise stationary sample where the retrospective changes are to be sought}

\item{minimum_distance}{A real number between 0 and 1 corresponding to a lower-bound on the minimum normalized length of the stationary segments (as percentage of total sample length)}
}
\value{
The list of changepoints in order of score
}
\description{
Returns the position of changepoints in the sequence
}
